BUGFIX. FIX FLOATING POINT ROUNDING AND TESTS
- Minor bugfixes, removal of empty lines and other less significant
stuff.

- In a few places in the code, in order to compensate for situations
where floating point arithmetic ended up in a number x slightly larger
than 1, the function np.cos(x) would return np.nan which was previously
fixed with some additional logic, but the logic itself was faulty due to
it also rounding numbers slighlty smaller than 1 which resultet in a
loss of precision. In this commit, the issue is resolved.